+++ WebApplication/iFlog/Controls/OnlineAdView.ascx.vb
@@ -77,21 +77,8 @@ Partial Public Class OnlineAdView
                 lblContactName.Text = .ContactName
             End If
 
-            If .ContactValue = "" Or .ContactName = "Private" Then
-                objContactDetail.Visible = False
-            Else
-                If .ContactType.ToLower = "email" Then
-                    Dim r As Regex = New Regex("\b[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}\b", RegexOptions.IgnoreCase)
-                    Dim m As Match = r.Match(.ContactValue)
-                    If m.Success Then
-                        litContactDetails.Text = String.Format("<a href=""mailto:{0}"">{1}</a>", .ContactValue, .ContactValue)
-                    Else
-                        litContactDetails.Text = .ContactValue
-                    End If
-                Else
-                    litContactDetails.Text = .ContactValue
-                End If
-            End If
+            litContactEmail.Text = .ContactEmail
+            litContactPhone.Text = .ContactPhone
 
             objPrice.Visible = onlineAd.Price > 0
             lblPrice.Text = String.Format("{0:C}", .Price)