+++ WebApplication/iFlog/Controls/OnlineAdViewWithDiv.ascx.designer.vb
@@ -159,22 +159,40 @@ Namespace Controls
         Protected WithEvents lblContactName As Global.System.Web.UI.WebControls.Label
 
         '''<summary>
-        '''objContactDetail control.
+        '''objContactPhone control.
         '''</summary>
         '''<remarks>
         '''Auto-generated field.
         '''To modify move field declaration from designer file to code-behind file.
         '''</remarks>
-        Protected WithEvents objContactDetail As Global.System.Web.UI.HtmlControls.HtmlTableRow
+        Protected WithEvents objContactPhone As Global.System.Web.UI.HtmlControls.HtmlTableRow
 
         '''<summary>
-        '''litContactDetails control.
+        '''litContactPhone control.
         '''</summary>
         '''<remarks>
         '''Auto-generated field.
         '''To modify move field declaration from designer file to code-behind file.
         '''</remarks>
-        Protected WithEvents litContactDetails As Global.System.Web.UI.WebControls.Literal
+        Protected WithEvents litContactPhone As Global.System.Web.UI.WebControls.Literal
+
+        '''<summary>
+        '''objContactEmail control.
+        '''</summary>
+        '''<remarks>
+        '''Auto-generated field.
+        '''To modify move field declaration from designer file to code-behind file.
+        '''</remarks>
+        Protected WithEvents objContactEmail As Global.System.Web.UI.HtmlControls.HtmlTableRow
+
+        '''<summary>
+        '''litContactEmail control.
+        '''</summary>
+        '''<remarks>
+        '''Auto-generated field.
+        '''To modify move field declaration from designer file to code-behind file.
+        '''</remarks>
+        Protected WithEvents litContactEmail As Global.System.Web.UI.WebControls.Literal
 
         '''<summary>
         '''lblDatePosted control.